HACP: Allow objectives to be written using PutParam request
To add support for objectives, we change the interface to
hacp-put-param-request to take an assoc of HACP values instead of
keyword arguments.  This allows us to write arbitrary fields to the
server, including the HACP objective fields.  And this, in turn, means
we can save our entire program state over HACP with only a single
network request.