HACP: Use proper "percent encoding" for keys, values and data
Along the way, we also update our percent-encoding routines to know
about the official four "unreserved characters" from RFC 3986, which
should generally not be escaped ("_-~.").

Note that we do not percent-encode the "~" character, even though the
HACP standard thinks it is a reserved character when encoding URLs.
I've arbitrarily decided to go with the modern specification of URL
encoding for now, and not carry around a second, HACP-specific
version of those rules.  We can fix this later if it somehow becomes
necessary.