HACP: Rewrite high-level client to run asynchronously
This patch converts the high-level HACP client from a handful of
functions using global variables into an object with a state machine
that runs in the background.

An attempt has been made to write real unit tests for the asynchrnous
behavior, but we'll have to see whether these tests are actually
sufficient.