UpdateInstaller: Use boost:unordered_set, _map, and _multimap
Rather than storing our FileSet entries in a std::vector, we should
store them in an unordered set data structure, as this allows quicker
lookup to see if we have a matching entry.

This patch moves us from a std::vector of entries, to a
boost::unordered_set.  We get rid of the map from filenames to entries,
as that was simply used to perform the set membership operations, which
our EntrySet can now perform.  We also replace uses of stdext::hash_map
and _multimap with boost::unordered_map and _multimap, because they are
better documented and are much closer to the standard interfaces provided
by TR1, if we decide to upgrade to a compiler that supports that.

This patch also adds an unordered set of digests to FileSet, which we will
need for computing our moves and copies from the existing tree to the
pool.