UpdateInstaller: Implement file moves, deletes, and copies via the pool
This patch implements the file move and copy code documented in
todo/updater-update.txt.  Because we are going to be updating several
gigabytes worth of data, we don't want to make any extra copies of that
data if we can avoid it.  We avoid this by being tracking exactly which
files we need where, and moving or copying them as necessary.  See the
notes in todo/updater-update.txt for more details.