Remove border from stage in full-screen mode
We've been seeing a border around the stage in full-screen mode ever
since we added wxAUI support.  This patch overrides wxAuiDefaultDockArt
and removes the border.

Note that we really ought to remove this border only when we're in
full-screen mode, and that this patch causes some graphical artifacts in
editing mode.  But we're about to start a beta round, and I want to do
something simple and foolproof.