FileSystem: Simplify paths containing . and .. when setting base dirs
Commit b77b8608, upgrading Boost, managed to break media in Halyard
programs when run in command line mode or runtime mode (specifically,
when run in command line mode or runtime mode with a path containing a
"." component).  This happens because we include a "." component in
the path when running in those modes, when launched from our shortcut
or our "rake halyard:test" task.  Previous versions of Boost stripped
that "." out when turning it into an abosolute path, but the upgraded
Boost did not.  QuickTime does not seem to be capable of handling
paths that contain "." in them, and so it would error out when we
tried to play media through QuickTime.

This patch introduces a Simplify operation, which will simplify a path
by stripping out path components consisting entirely of ".", and also
walking up the hiererchy by stripping out parent path components when
encountering a ".." component.  We apply this operation when
constructing our base paths, to ensure that we won't get values that
certain APIs can't handle.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?17655