Updater: Fix path to re-launch executable
When we did the re-organization of Halyard to move the engine into
engine/win32, we neglected to update the path that Halyard passes to
the updater in order for the updater to re-launch Halyard.  This patch
updates that path, to explicitly point to engine/win32.

Note that that if this were to run on another platform, or we re-organize
the layout again, we will need to re-address this issue; this is a
stopgap measure, but it is the only one that's reasonable to apply
before we actually do any work on generalizing this.