Updater: Fix performance regression in download-update
When modifying the update downloader to only download a file if a file
with the same digest doesn't already exist on disk, I had introduced a
serious performance regression.  In order to check if a digest is
present on disk, I had been reparsing the entire set of manifest files
once for each digest I needed to check, and then doing a linear scan
of them.  Needless to say, while this didn't affect the unit tests
much, for large programs like VTRA this introduced a long pause before
downloading updates could begin.

This patch fixes this problem by memoizing the result of parsing
the manifests in the existing directory, as well as memoizing a
hash table indicating if digests are present on the user's system.
The values memoized depend on the currently active *updater* (which
can change between unit tests), so they will re-compute the thunk if
*updater* changes.