Updater: Extract superclass of UpdateInstallerTest
In order to be able to write more tests against a more complicated
fixture, without breaking existing tests, I extract a superclass from
UpdateInstallerTest which includes the code that "loads" a fixture
(by making a temporary copy of it and cding to that) and other
support routines that are independent of the fixture.

Extracting a superclass from Test::Unit::TestCase is actually fairly
difficult, as Test::Unit::TestCase by default tries to run tests
out of every class that inherits from it, and provides no way to
mark a class as being used simply as a superclass; it seems that
the preferred way of doing inheritance is with mixins, but that
doesn't work if you want to use class variables to pass some values
to your superclass.  So to workaround this, I simply define a
UpdateInstallerTest::suite method that returns an empty test suite
if called directly on this superclass.