Updater: Delete vestigial files in script and collect directories
In previous versions of the updater, we did not delete files that were
already installed but not needed in the new version.  This didn't cause
us any problems, because we didn't delete files very often, and extra
files usually aren't harmful.

In the wake of the big reorg for the 0.5 series of engines, however,
we are more likely to be left with vestigial files and directories.
We have also discovered that extra empty directories in the wrong place
can screw up PLT's loader.  So, we're going to need to clean up any
extra files and directories within any of our load paths; scripts,
collects, engine/win32/collects, and engine/win32/plt.

This patch adds just the code for deleting extra .ss, .zo, and .dep
files from any of the directories in the load path.  We still need
to keep track of what directories are going to be empty, and delete
those as well.