Updater: Delete vestigial directories in script and collect dirs
In addition do deleting old .zo, .ss, and .dep files, it is necessary for
us to clean up any unwanted directories.  Extra directories in our
search path can sometimes cause problems for PLT's module loader.  This
patch tracks which directories we are going to want in the updated program,
and tries to delete any other directories.  If any unkown files are found
in these empty directories, however, they cannot be deleted.  This is a
bit more forgiving than the last patch, which would cancel the update
whenever it found an unknown file anywhere within the scripts or collects
directory; with this patch, we assume that unknown files without a known
bad file extension are innocuous, unless they prevent us from deleting
an empty directory.