Updater: Downcase file and directory names when doing cleanup.
While doing the cleanup of our script and collect directories, we are
reading filenames off of the disk and comparing them to lists of filenames
from our base and update manifests.  Because VFAT and NTFS are case
insensitive, case preserving, we need to do this comparison
case-insensitively, or else we may wind up with a discrepancy between
what we see on disk and what our manifest says should be present.

This patch modifies the compution of a filename set to lowercase
all filenames before inserting them.  In order to make this clear,
I rename FileSet::FilenameSet to FileSet::LowercaseFilenameSet,
with equivalent renaming of variables.  When iterating over the
directory and computing our filenames relative to the program root
that we are going to check against our manifests, we also need to
downcase the filenames to check them against our set.