Updater: Add logging when marking updates as impossible
In the past, when we have detected that we cannot update, we have simply
silently relaunched the program without doing anything.  This did not
occur often, as it would only occur if for some reason the user couldn't
write to the destination directory or was missing a file in the pool.
This is clearly not a very good response to an error, as the update will
keep mysteriously failing without any indication of why.

We've added a few other ways for updates to mysteriously fail with the
smarter updater code and cleanup code.  Because we have more paths to
failing, any failures we do get will be that much more mysterious.  This
patch improves that situation a bit, by adding logging to all of the new
code paths in which we mark updates as impossible.

There are also a few notes added to the updater todo about a possible
better mechanism we could use when we encounter such errors, but we
don't have the time to implement that at this moment.