Updater: Fix bug in which we would try to delete non-empty directories
When implementing the cleanup code that deleted unwanted files and empty
directories, we need to keep track of which directories do contain files
too keep, and not delete those.  When assembling that list, however, I
had only added the parent directories of files that exist in our manifest,
not any other ancestors.  This meant that if you had a deeply nested
set of directories in which only the innermost contained any files,
we would try to delete the outer, non-empty directories, which would
cause the update to die.  This patch adds tests demonstrating the problem,
and a fix which walks up the pathname for any files to keep and adds
all ancestor directories to the list of directories to keep as well.