Updater: Rename directories that differ only in case
When installing over an old version of the program, we may wind up
with files and directories on disk which differ in case from what
their new name would be.  This can cause all kinds of problems in
PLT's module loader, which likes to compare pathnames case-
sensitively.  Our updater would previously do nothing about directories
with the wrong case, as it would see a change from FOO/BAR to foo/bar
as a move of that file, which because we're on a case insensitive file
system, would leave the file as FOO/bar.

This patch adds a new phase of the update install that cleans up all
pathnames that differ in case from the existing manifest to the
update manifest, by walking up the pathnames to their ancestor path
and issueing move commands for each component along the way.