Updater: Case rename directories even when all files within them change
In code review, Eric noticed a bug in which a directory which changed
case contained no files in common between the existing manifest and the
update manifest.  In this case, we would never compute the renames
for the case change, and the directory would be left as is.

This patch rewrites the case rename code to deal with this bug.  Indtead
of comparing filenames between the old and new manifest, and adding
rename commands for their ancestor directories, we operate on the
directories themselves.  We compute a map from lowercased directory
names to mixed case directory names, and for each directory in the
updated tree, we check if it is present in the existing and whether
it mismatches in case, creating a rename command if it does.