Updater: Look for executable in various places if we don't find it
When calling out to our update installer, our programs pass in the name
of the executable and arguments that the update installer should
call to re-launch the program once it is done (though this works only
on Windows XP, as we can't shed privileges on Vista).  When we
reorganized Halyard, however, the executable moved from Tamale.exe
to engine/win32/Halyard.exe.  This patch adds some code to check if the
requested executable exists, and if not, look in various paths
for an executable to launch.