Updater: Code review cleanup
This patch addresses a few issue Eric noticed in code review.  When
constructing lists of directories on disk, we return them to the calling
function as a return value.  He mentioned that this might be expensive,
so I've added references which indicate that Visual C++ performs
named return value optimization, so we shouldn't need to worry about
this.

This patch also limits the list of places we look for an executable
when trying to relaunch after an update, and stops looking as soon
as we find one.