Updater: Add LnkTrampoline project to redirect user with old shortcuts
When updating from SAE 1.0 to 1.5, users will have shortcuts that
point to Tamale.exe.  Our engine has been renamed to Halyard.exe,
however.  In order to solve this, we add a new LnkTrampoline project,
which builds a stub Tamale.exe which, when invoked, simply
ShellExecutes engine/win32/Halyard.exe with the same arguments
that it was called with.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?17956