Updater: LnkTrampoline: Use backslashes when launching engine
Apparently, while forward slashes work in ShellExecute on XP (and just
about everywhere else I've tried them), they don't work on Vista or
newer.  So, we switch the slashes to backslashes, as well as adding
some error reporting machinery that was helpful in debugging the
problem.  This has been tested on Vista and XP, but not yet tested on
Windows 7, so hopefully the extra error reporting will help us better
track down any issues that may come up.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?17956