data-file.ss: Use Tamale user data directory if specified and exists
When migrating from Tamale to Halyard, we may wish to continue to use
a users old pref files.  The simplest way to do so is to simply change
which directory we load their user prefs out of; this way, we don't
have to worry about any sort of file copying issues, and even if they
downgrade again, they will still have their new progress.

In order to enable loading of user prefs out of the Tamale directory,
add a file named config/USE-TAMALE-DIRECTORY-FOR-USER-PREFS

Bug: http://iml.dartmouth.edu/fogbugz/default.php?17973