Use Tamale user data directory for all script user data access
In the previous patch, we had redirected the user data path used by
data-file.ss to the old Tamale user data directory if the program
requested this and the Tamale user data directory existed.  Some code
in IML libs was accessing (script-user-data-directory) directly to find
user data files, so it wasn't picking up that redirection.

This patch moves that redirection from data-file.ss to
(script-user-data-directory) in order to give a consistent view of
where user data is located.