Correctly enable/disable accelerators in full-screen mode
The problem: wxFrameBase::UpdateUI no longer does anything useful with
menu items, as described in the comments added by this patch.  So we
leave in our call to UpdateUI (just in case it turns out to be doing
something else useful that we _don't_ know about), and we add extra code
which tries to replicate the actual menu->UpdateUI(...) calls made when
we're running in AUTHORING mode.  In personal testing, it appears to
correctly disable Control-O in halyard/test once a program has been
loaded.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?17748