LICENSE.txt: Convert line endings to CRLF for use on Windows
This file gets linked to from the Start menu on Windows when we build
installers, and it will generally get opened in Notepad, which doesn't
support LF line endings.  So let's just convert this whole file to
CRLF format for now, and not try to come up with some way to
automatically convert it when making a Windows build.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?14218