Add %card% .focus! method
We've received bug reports about the stage not being properly focused
on startup, which prevents users from using custom key bindings right
after the program has started.  This tends to affect demo users and
instructors, who know various semi-secret key codes to activate a
special program mode.

I tried focusing the stage on startup immediately after calling Show()
on the Stage, but it appears that we either fail to gain the focus in
the first place, or we lose it at some point later on.

However, this %card% .focus! method will allow cards early in the
startup process to properly focus themselves.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?16994